Streamline inventory item creation and clean up Tile objects
Previous to this change, inventory item creation was cheating, whereby a Tile game object was prebuilt & hidden, then duplicated and activated on command. This has been replaced with an interface in TileData to spawn the Game Object.

Integral to these changes is Tile cleanup. TileData is now being used as the Model for the Tile Game Object. The Tile component we add to the Game Object is now treated as the Controller. The previously separated objects of Controller/Component have been merged into a single object.

Going forward, all Component/Controller separations will be removed in favour of this new pattern, as changes to those updates occur.

See https://github.com/JerkyTreats/EndlessHorizons/issues/9