Allow Blueprint.Load to spawn Ship Components loaded from external JSON file
The spawn is using a `switch` on a ship component enum to determine the correct type to create. This may be target for future refactor

Ship Components need to save the file name of the external JSON file they are loaded from. The filename acts a key for the ComponentRepository

See https://github.com/JerkyTreats/EndlessHorizons/issues/11