Add Logging System
Adds initial logging system. Features a singleton `Logger` class which governs if/how a message is logged, and a `Log` class managing what gets logged.

The logging system features a `GameAreas` enum that governs if a message will be logged. Each class can specify the GameArea they affect, and logging of GameAreas can be turned on and off at will via the Logger instance. The result is a fairly powerful logging system to log specific areas of code as necessary.