Refactor SharedEdge merge function
Merging two SharedEdges requires that the First/Last value in the SharedEdge matches the Last/First value in the SharedEdge to merge.

Before, we had this conditional in the SharedEdgeFactory, with two methods in SharedEdge for merging either at the beginning or end of the existing SharedEdge.

That conditional is now moved into a single Merge method in SharedEdge.